Float gyro offset

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/sensors/gyro/Gyro.java
@@ -9,5 +9,5 @@ public interface Gyro extends Available {
 
     float getRaw();
 
-    void setOffset(int offset);
+    void setOffset(float offset);
 }
